Remove SID from TUS fingerprint
This enables to resume an upload even if the browser tab has been
closed. Although bad things can happen if a user tries to upload two
different files with identical file name, file size and file last
modified time stamp. Including the MD5 sum of the file should resolve
this, but not yet done. Also if the solution might work, it seems to be
a bit weird. For a new upload a new sid is generated, but the old one is
found via fingerprint and used as endpoint and resume works. Probably it
would better not to generate a new sid in this case.

See also psi-4ward/psitransfer/issues#57.